backend/requirements.txt
# ========================================
# Medical Billing Assistant Backend
# Python Dependencies
# ========================================

# ===== Core Web Framework =====
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# ===== AI & Language Models =====
openai>=1.22.0,<2.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-core>=0.3.34
langchain-community>=0.0.2
langchain-pinecone>=0.0.1

# ===== Vector Database =====
pinecone-client>=2.2.4

# ===== Data Validation & API =====
pydantic>=2.7.0
python-multipart>=0.0.6

# ===== Real-time Communication =====
websockets>=11.0
websocket-client>=1.6.4

# ===== HTTP & File Handling =====
requests>=2.31.0
aiofiles>=23.2.1

# ===== Authentication & Security =====
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ===== Testing =====
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# ===== PDF Processing =====
pdfplumber>=0.10.0
PyPDF2>=3.0.1

# ===== Data Processing =====
pandas>=2.0.0
numpy>=1.25.0

# ===== Audio & Speech =====
pyttsx3>=2.90

# ===== Utilities =====
# Note: uuid, asyncio, json, os, sys, base64, re, csv, logging are standard library modules